--- lib/src/features/calendar/view/forms/components/event_color_picker_item.dart
@@ -1,47 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_colorpicker/flutter_colorpicker.dart';
-
-class EventColorPickerItem extends StatefulWidget {
-  final Color color;
-  final bool isCurrentColor;
-  final VoidCallback changeColor;
-
-  const EventColorPickerItem(
-      {super.key,
-      required this.color,
-      required this.isCurrentColor,
-      required this.changeColor});
-
-  @override
-  State<EventColorPickerItem> createState() => _EventColorPickerItemState();
-}
-
-class _EventColorPickerItemState extends State<EventColorPickerItem> {
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(5.0),
-      child: DecoratedBox(
-        decoration: BoxDecoration(
-          shape: BoxShape.circle,
-          color: widget.color,
-        ),
-        child: Material(
-          color: Colors.transparent,
-          child: InkWell(
-            onTap: widget.changeColor,
-            borderRadius: const BorderRadius.all(Radius.circular(50)),
-            child: AnimatedOpacity(
-              duration: const Duration(milliseconds: 210),
-              opacity: widget.isCurrentColor ? 1 : 0,
-              child: Icon(Icons.done,
-                  color: useWhiteForeground(widget.color)
-                      ? Colors.white
-                      : Colors.black),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
